fix: csrf and signedIn in local storage deleted on refresh
Problem:
Since the response api call is attached to every api request on error,
when there's two concurrent requests that fails, they will both call
the refresh endpoint. In which, this will return a success response
to one, then a failed response on the other due to the tokens being
refresh earlier on. So on failed refresh, the csrf and signedIn in
local storage gets cleared in which triggers the user to logout.

Solution:
Only add the refresh api call on "/auth" route.